Added the following operators: +, -, /, *, +=, -=, /=, *=
The operators that were added are not usable with the types supported by
lua_value. The following operators +, -, /, * have return type same as the  other
operant, e.g lua_value_ref + int returns int. The assigment operators return lua_value_ref& as
per convertion. For all operators there are no unit test, this is because it's not possible to test them fully
without creating dependencies between the various unit test. The operators already listed are deleted when the rhs
is of type lua_value_ref, this decision has been made because; the lua c api does not permit to perform the these operation
directly on the lua stack, this means that we have to first transform them into C++ types. While is possible to check the type of the lua variable,
it's not possible to decide for the user how a variable should be casted, for example. Let's say we have the following line of code
lua_script["a"]+lua_script["b"], and let's say "a" is an integer with value 1 and "b" is a number with value 1.2, it's not possible for us
to decide for the user whether "a" should be casted as short,int or long as it's not possible for us to decide whether "b" should be caster as float or double.
For this reason the operators accepting const lua_value_ref&  are marked as deleted, in order to perform the listed operators we first have to assign one of the lua_value_ref variables
to a c++ variable with explicit type and only then perform the operator.